feat(core): deduplicate log targets
Previously, one log targets can belong to many log containers. Consider this in Javasript:

```
const foo = {
  test: () => {
    const bar = "bar"
  }
}
```

If you log the variable `bar`, it will result in two lines of logs, one for the inner function container, one for the outer declaration container.

This PR fixes that by picking the deepest container when there are more than one. We assume that users want the smallest scope possible.