feat!(watcher): update the sources configuration
Previously, we specified `sources` for watcher as an array and use name as the identifier. However, this restricts us from having multiple sources with the same name. Also, we rely on source name being unique, but nowhere does it being enforced in the source code.
This PR updates the `sources` config to a table instead, with keys are source id. We use source id as identifier. And since they are table keys, they are guaranteed to be unique.